Change the tetromino spawn order
The spawn method follows the "Random Generator" algorithm in
the Tetris Guidelines. The shapes are put into a pool and randomized.
When all shapes in the pool have been used up, the process is repeated.

In order to allow for a preview system, a second pool is used which is
one step ahead of the main pool.