Put global window variables in a struct
There is a mix of platform specific and independent code, so the struct
currently has two backbuffer representations. The width and height can
also maybe be made into functions which take their value from the
backbuffer so there is no need to update them every time the window is
resized.