Make all colors use an RGBA struct
Previously there were both RGB and RGBA structs and some functions took
an RGB with an additional alpha parameter. Having just a single RGBA
struct should simplify things and make them more consistent.
Additionally, the RGB/A structs were typedef'd to V3/4<int> which would
then be the same type as other usages of V3/4<int>. It makes for a messy
and loose type system which I want to tighten.