Bunch of stuff related to the Positive class
Make members in BackBuffer use the Positive class. This requires some
casts in different places.

Use the point_is_in_rect() function for the check in draw_hollow_square
whether or not the pixel is part of the border or not.

Comment out the draw_image function since it's not being used currently
anyway.

Make the signed constructors of PositiveGeneric explicit.

Make the unsigned constructors of PositiveGeneric constexpr and do a
constexpr check to see if the constructors type is larger than T, in
which case an assert is used to make sure it fits.

Add comparison operators to PositiveGeneric.

Remove the RGBA::Unchecked struct. Since the unsigned constructors are
now constexpr, the Unchecked struct has lost its purpose.