Widen the u8 before multiplying to silence overflow warning
If the values of rows and columns are unknown, the multiplication could
overflow the int before being upcasted to a size_t. By widening one of
the factors beforehand, this potential problem is avoided.

Of course, this wouldn't be a problem if the static analyzer realized
that both factors are constexpr values and actually verified that the
multiplication was safe :)